requirements.txt
flask==3.0.2
azure-cognitiveservices-speech==1.34.0
pyyaml==6.0.1 